requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.6
dj-database-url==3.1.2
Django==5.2.12
gunicorn==25.2.0
packaging==26.0
psycopg==3.3.3
psycopg-binary==3.3.3
pycparser==3.0
sqlparse==0.5.5
whitenoise==6.12.0
djangorestframework==3.16.1
